requirements.txt
google-generativeai
flask
werkzeug
ffmpeg-python
python-dotenv
fpdf2
google-auth
requests
